+++ lib/extglob/compiler.js
@@ -8,13 +8,16 @@ var utils = require('../utils');
  * @param {Object} `options`
  */
 
-function Compiler(options, state) {
+function Compiler(options) {
   debug('compiler', __filename);
   this.options = options || {};
-  this.state = state || {};
   this.compilers = {};
   this.output = '';
 }
+
+/**
+ * Prototype methods
+ */
 
 Compiler.prototype = {
