+++ lib/extglob/parser.js
@@ -234,8 +234,18 @@ Parser.prototype = {
     return this;
   },
 
+  /**
+   * Create a parser with open and close for parens,
+   * brackets or braces
+   */
+
   pair: function(type, openRegex, closeRegex) {
-    this.set(`${type}.open`, function() {
+
+    /**
+     * Open
+     */
+
+    this.set(type + '.open', function() {
       var parsed = this.parsed;
       var pos = this.position();
       var m = this.match(openRegex);
@@ -245,7 +255,7 @@ Parser.prototype = {
       var prev = this.prev();
       this.specialChars = true;
       var open = pos({
-        type: `${type}.open`,
+        type: type + '.open',
         val: val
       });
 
@@ -259,27 +269,28 @@ Parser.prototype = {
 
       utils.define(node, 'parent', prev);
       utils.define(open, 'parent', node);
-      if (this.specialChars) {
-        node.specialChars = true;
-      }
 
       this.push(type, node);
       prev.nodes.push(node);
       return node;
     });
 
-    this.set(`${type}.close`, function() {
+    /**
+     * Close
+     */
+
+    this.set(type + '.close', function() {
       var pos = this.position();
       var m = this.match(closeRegex);
       if (!m) return;
 
       var open = this.pop(type);
       if (this.isType(open, type)) {
-        throw new Error(`missing opening "${type}"`);
+        throw new Error('missing opening "' + type + '"');
       }
 
       var node = pos({
-        type: `${type}.close`,
+        type: type + '.close',
         rest: this.input,
         suffix: m[1],
         val: m[0],
@@ -335,13 +346,13 @@ Parser.prototype = {
       var prev = this.input;
       this.next();
       if (this.input && prev === this.input) {
-        throw new Error(`no parsers registered for: "${this.input.charAt(0)}"`);
+        throw new Error('no parsers registered for: "' + this.input.charAt(0) + '"');
       }
     }
 
     if (this.stack.length && this.options.strict) {
       var node = this.stack.pop();
-      throw this.error(`missing opening ${node.type}: "${this.orig}"`)
+      throw this.error('missing opening ' + node.type + ': "' + this.orig + '"')
     }
 
     this.eos();
