+++ lib/extglob/position.js
