+++ lib/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var Compiler = require('./snapdragon/compiler');
-var Parser = require('./snapdragon/parser');
+var Compiler = require('./extglob/compiler');
+var Parser = require('./extglob/parser');
 var utils = require('./utils');
 
 function Extglob(options) {
@@ -28,6 +28,8 @@ Extglob.prototype.parse = function(str, options) {
     }
   }
 
+  // add non-enumerable parser reference
+  utils.define(parsed, 'parser', this.parser);
   return parsed;
 };
