+++ lib/utils.js
@@ -17,5 +17,5 @@ exports.last = function(arr) {
 };
 
 exports.not = function(str) {
-  return `^((?!(?:${str})).)*`;
+  return '^((?!(?:' + str + ')).)*';
 };
